pyproject.toml
[project]
name = "frugalai-ml-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "datasets>=3.6.0",
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-cloud-storage>=3.1.0",
    "pandas>=2.3.0",
    "peft>=0.15.2",
    "scikit-learn>=1.6.1",
    "torch>=2.7.1",
    "trl>=0.18.1",
]
